pyproject.toml
[tool.poetry]
name = "pdf-rag-template-v2"
version = "0.1.0"
description = ""
authors = ["ivan-pua <qieshang@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
langchain-community = "^0.3.3"
langchain-openai = "^0.2.3"
pypdf = "^5.0.1"
faiss-cpu = "^1.9.0"
streamlit = "^1.39.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
